requirements.txt
mypy
black
isort
discord.py
discord.py-stubs
